[PATCH 1/4] libsystemd-shared/log: ignore prohibit_ipc flag in container
*prohibit_ipc* flag is used to deny inter-process communication
to logging methods during early system startup (before journald is running)
and during system shutdown (when journald might not be running anymore).

in a containerized systemd running in a systemd host,
verify if running in container when attempting to log_open()
and if so, ignore the prohibit_ipc flag with the expectation that
connection is established to host namespace journald.

assumes host namespace journald socket is exposed (bind mounted)
to the container at the expected path (of /run/systemd/journal).

functionality enabled by defining HOSTONLY_JOURNALD.